Exceptions with context and logouts with cookie

+++ src/Auth/Application/UseCase/LogoutAll.php
@@ -1,9 +1,11 @@
 <?php
 
 namespace App\Auth\Application\UseCase;
-use App\Auth\Domain\ValueObject\UserId;
 use App\Auth\Domain\Repository\RefreshTokenRepositoryInterface;
 use App\Shared\Application\Port\CookieManagerInterface;
+use App\Auth\Domain\Exception\InvalidTokenException;
+use App\Auth\Domain\ValueObject\TokenValue;
+
 
 
 final class LogoutAll {
@@ -12,8 +14,12 @@ final class LogoutAll {
         private readonly CookieManagerInterface $cookieManager
     ){}
 
-    public function logoutAll(string $userId):void {
-        $this->tokenRefreshRepository->deleteAll(UserId::fromString($userId));
+    public function logoutAll():void {
+        $refreshTokenValue = $this->cookieManager->get('refreshTokenJKApp');
+        if(!$refreshTokenValue) throw new InvalidTokenException();
+        $refreshToken = $this->tokenRefreshRepository->findByTokenValue(TokenValue::fromString($refreshTokenValue));
+        if(!$refreshToken) throw new InvalidTokenException();
+        $this->tokenRefreshRepository->deleteAll($refreshToken->userId());
         $this->cookieManager->delete('refreshTokenJKApp');
     }
 }
